requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pyswisseph==2.10.3.2
python-dotenv==1.0.1
pydantic==2.10.6
pytest==8.3.4
httpx==0.28.1
